code

+++ t5/main.go
@@ -1,6 +1,6 @@
 package main
 
-func Add(a int64, b int64) int64 {
+func Add(a int32, b int32) int32 {
 	return a + b
 }
